Move info command to separate file
